requirements.txt
flask
mysql-connector-python
werkzeug
openpyxl
reportlab
